DO NOT MERGE - broken gson class
This do not yet reproduce issue.
The visible failure during class loading is the one I see, hower the
reproducible steps are not the same.

In my original scenario, I obtain Gson class from running jvm,  jdis -g it, and
jasm it back.  It ends as here pushed GsonWeird.class , which is
malformed (without -g it works fine)

However in this test, the jdis -g of GsonOrig.class, jasm and class laod works fine (?!?).
The GsonWeird.class which fails in this tests (with both -g and without)
is the corrupted output from my original scenario.
the GsonOrig.class is the freshly obtained class from my original
scenario