Temporary workaround around tests being order-sensitive
As -g is now making some disassmebld code not asemble-able back, and
order of tests is not deterministic, and Options are static, thus if -g
is now set in some test, it is not unset in other tests.

Real fix is to move Options out of static context to context of
environment as it is done for jcoder.