made maven resource plugin expclicit
That is still sufferign the https://issues.apache.org/jira/browse/MRESOURCES-237
so the filtering was done. Fitlering is excluding proeprties (as they
claim they are ascii) and is adding .class sufix for test classes.
Another solution is to use mrp 2.6 which do not suffer the symlinks
issue

In both cases, the tests still fail in maven runs, although rummor is,
manually it passes (todo, verify)